refactor(db): reorganize file system
- Move migrations to `src/` (it's part of production code)
- Refactor tests (use helpers)
- Etc.
- Reflect changes on config
