--- packages/db/.gitignore
@@ -1,2 +1,0 @@
-tmp
-test/tmp
