+++ packages/db/knexfile.js
@@ -2,25 +2,27 @@ module.exports = {
   development: {
     client: 'sqlite3',
     connection: {
-      filename: './tmp/sqlite3'
+      filename: './tmp/dev.sqlite3'
     },
     migrations: {
-      directory: './data/migrations'
+      directory: './src/migrations'
     },
     seeds: {
       directory: './data/seeds'
-    }
+    },
+    useNullAsDefault: true
   },
   test: {
     client: 'sqlite3',
     connection: {
-      filename: './test/tmp/sqlite3'
+      filename: './tmp/test.sqlite3'
     },
     migrations: {
-      directory: './data/migrations'
+      directory: './src/migrations'
     },
     seeds: {
       directory: './test/data/seeds'
-    }
+    },
+    useNullAsDefault: true
   }
 }
