+++ packages/db/test/data/seeds/01-users.js
@@ -1,4 +1,4 @@
-const { users } = require('./snapshot.json')
+const { users } = require('../snapshot.json')
 
 exports.seed = async knex => {
   // Deletes ALL existing entries
