+++ packages/db/test/helpers/hooks.js
@@ -1,0 +1,8 @@
+module.exports.up = knex => async () => {
+  await knex.migrate.latest()
+  await knex.seed.run()
+}
+
+module.exports.down = knex => async () => {
+  await knex.migrate.rollback()
+}
