+++ packages/db/test/helpers/knex.js
@@ -1,0 +1,11 @@
+const Knex = require('knex')
+
+const settingsMap = require('../../knexfile')
+
+//
+
+const ENV = 'test'
+
+// init
+
+module.exports = Knex(settingsMap[ENV])
